Move Rayon-related functionality from rten-base to rten-parallel
Refactor the rten-base crate so that it no longer depends on Rayon. Instead the
parts that depend on Rayon have been moved into a new rten-parallel crate. This
enables crates such as rten-model-file and rten-vecmath, which depend on
rten-base but don't use the Rayon-related functionality, to build earlier.

In the process of doing this, the documentation for `ParIter` was improved to
include an end-to-end example.
