+++ rten-gemm/Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/robertknight/rten"
 [dependencies]
 rayon = { workspace = true }
 rten-base = { path = "../rten-base", version = "0.24.0" }
+rten-parallel = { path = "../rten-parallel", version = "0.24.0" }
 rten-simd = { path = "../rten-simd", version = "0.24.0" }
 rten-tensor = { path = "../rten-tensor", version = "0.24.0" }
