+++ rten-parallel/src/lib.rs
@@ -1,0 +1,6 @@
+//! Parallel-processing utilities.
+//!
+//! This crate provides common utilities related to multi-threaded parallelism,
+//! for use by the RTen crates.
+
+pub mod par_iter;
